[test-api-test-init] : initialisation de jest et mise en place de test de base
